@@ -2,15 +2,16 @@
   "name": "@cycjimmy/awesome-js-funcs",
   "version": "0.0.0-development",
   "description": "Personal collection of javascript functions often used.",
+  "type": "module",
   "main": "cjs/index.js",
   "module": "esm/index.js",
   "scripts": {
     "lint": "eslint -c .eslintrc .",
     "lint:fix": "eslint --fix -c .eslintrc .",
-    "babel": "npx babel src --out-dir dist/esm",
-    "transform2cjs": "cross-env NODE_ENV=cjs npx babel dist/esm --out-dir dist/cjs",
+    "babel": "npx babel src --out-dir dist/esm --out-file-extension .mjs",
+    "transform2cjs": "cross-env NODE_ENV=cjs npx babel dist/esm --out-dir dist/cjs --out-file-extension .cjs",
     "build": "npm run lint:fix && trash dist && npm run babel && npm run transform2cjs",
-    "package": "npm run build && node modifyPackage.mjs && gulp copy:package"
+    "package": "npm run build && node modifyPackage.js && gulp copy:package"
   },
   "repository": {
     "type": "git",
@@ -30,15 +31,15 @@
     "access": "public"
   },
   "devDependencies": {
-    "@babel/cli": "^7.16.0",
-    "@babel/core": "^7.16.0",
-    "@babel/preset-env": "^7.16.4",
-    "@cycjimmy/config-lib": "^1.8.0",
+    "@babel/cli": "^7.17.6",
+    "@babel/core": "^7.17.7",
+    "@babel/preset-env": "^7.16.11",
+    "@cycjimmy/config-lib": "^3.0.2",
     "cross-env": "^7.0.3",
-    "eslint": "^8.2.0",
+    "eslint": "^8.11.0",
     "eslint-config-airbnb-base": "^15.0.0",
-    "eslint-plugin-import": "^2.25.3",
+    "eslint-plugin-import": "^2.25.4",
     "gulp": "^4.0.2",
-    "trash-cli": "^4.0.0"
+    "trash-cli": "^5.0.0"
   }
 }
