@@ -1,4 +1,4 @@
-const gulp = require('gulp');
+import gulp from 'gulp';
 
 // copy
 gulp.task('copy:esm', (done) => {
