@@ -1,5 +1,5 @@
 /* eslint import/no-extraneous-dependencies: ["error", {"devDependencies": true}] */
-const makeConfig = require('@cycjimmy/config-lib/semanticRelease/15.x/makeConfig');
+const makeConfig = require('@cycjimmy/config-lib/cjs/semanticRelease/15.x/makeConfig.cjs').default;
 
 module.exports = makeConfig({
   exec: true,
